Break Calendar.tsx into separate components
