+++ app/(product)/components/ConfirmModal.tsx
@@ -1,0 +1,62 @@
+"use client";
+
+import React from "react";
+
+export default function ConfirmModal({
+  title,
+  description,
+  confirmText = "Confirm",
+  cancelText = "Cancel",
+  confirmVariant = "danger",
+  onCancel,
+  onConfirm,
+}: {
+  title: string;
+  description?: React.ReactNode;
+  confirmText?: string;
+  cancelText?: string;
+  confirmVariant?: "danger" | "success";
+  onCancel: () => void;
+  onConfirm: () => void | Promise<void>;
+}) {
+  const [busy, setBusy] = React.useState(false);
+  const handleConfirm = async () => {
+    try {
+      setBusy(true);
+      await onConfirm();
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const confirmClasses =
+    confirmVariant === "success"
+      ? "rounded-md bg-green-700 px-4 py-2 text-sm font-medium text-white hover:bg-green-800 disabled:opacity-50"
+      : "rounded-md bg-red-700 px-4 py-2 text-sm font-medium text-white hover:bg-red-800 disabled:opacity-50";
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50">
+      <div className="w-full max-w-md rounded-lg bg-white dark:bg-gray-900 p-6 shadow-xl border border-gray-200 dark:border-gray-800 text-gray-900 dark:text-gray-100">
+        <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-100">{title}</h2>
+        {description && (
+          <div className="mt-3 text-sm text-gray-700 dark:text-gray-300">{description}</div>
+        )}
+        <div className="mt-6 flex justify-end gap-3">
+          <button
+            className="rounded-md border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-900 text-gray-900 dark:text-gray-100 px-4 py-2 text-sm hover:bg-gray-100 dark:hover:bg-gray-800"
+            onClick={onCancel}
+            disabled={busy}
+          >
+            {cancelText}
+          </button>
+          <button
+            className={confirmClasses}
+            onClick={handleConfirm}
+            disabled={busy}
+          >
+            {busy ? "Please wait…" : confirmText}
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
